+++ NEWS.md
@@ -1,3 +1,10 @@
+# plinkQC 1.0.2
+## minor changes
+* Changed it so ancestry prediction function works with PLINK 1.9v files. There is
+now an additional parameter the user can use if the data is in PLINK 2.0 format
+* Added convert_to_vcf() function to convert vcf files to PLINK format 
+* Added parameter to ancestry function so user can use their own trained classifier
+
 # plinkQC 1.0.1
 ## minor changes
 * Fixed a bug so that ancestry prediction function works when there are phenotypes included in the input data.